+++ base_layer/core/src/mempool/unconfirmed_pool/mod.rs
@@ -23,9 +23,7 @@
 mod error;
 #[allow(clippy::module_inception)]
 mod unconfirmed_pool;
-mod unconfirmed_pool_storage;
 
 // Public re-exports
 pub use error::UnconfirmedPoolError;
 pub use unconfirmed_pool::{UnconfirmedPool, UnconfirmedPoolConfig};
-pub use unconfirmed_pool_storage::UnconfirmedPoolStorage;
